# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Conference Management Tool API",
        "description": "API documentation for the Conference Management Tool (CMT).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/reports/dashboard": {
            "get": {
                "tags": [
                    "Reporting"
                ],
                "summary": "Get dashboard summary",
                "description": "Returns summary statistics for users, conferences, submissions, reviews and registrations.",
                "operationId": "67c2d516c6890651d0c1ee6b7738ccf6",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reports/submissions": {
            "get": {
                "tags": [
                    "Reporting"
                ],
                "summary": "Get submission statistics",
                "description": "Returns submission statistics grouped by status with optional conference, status and date range filters.",
                "operationId": "5c03bc8c4e8ff1bd55149241b8b25835",
                "parameters": [
                    {
                        "name": "conference_id",
                        "in": "query",
                        "description": "Filter submissions by conference ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter submissions by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "under_review",
                                "accepted",
                                "rejected",
                                "revision_requested"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Return submissions created on or after this date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Return submissions created on or before this date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission statistics retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reports/reviews": {
            "get": {
                "tags": [
                    "Reporting"
                ],
                "summary": "Get review statistics",
                "description": "Returns review progress and statistics including total, submitted, pending and locked reviews, average score and recommendation distribution.",
                "operationId": "ef1357fc35ae36c83c98838b21cabe83",
                "parameters": [
                    {
                        "name": "conference_id",
                        "in": "query",
                        "description": "Filter reviews by conference ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Include reviews created on or after this date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Include reviews created on or before this date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review statistics retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reports/registrations": {
            "get": {
                "tags": [
                    "Reporting"
                ],
                "summary": "Get registration statistics",
                "description": "Returns registration statistics with optional conference, status and date filters.",
                "operationId": "53c02b10b43b8ff491b7e15e10c47540",
                "parameters": [
                    {
                        "name": "conference_id",
                        "in": "query",
                        "description": "Filter registrations by conference ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter registrations by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "registered",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Return registrations created on or after this date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Return registrations created on or before this date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Registration statistics retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reports/conferences": {
            "get": {
                "tags": [
                    "Reporting"
                ],
                "summary": "Get conference statistics",
                "description": "Returns statistics for conferences, including submission status, format and category.",
                "operationId": "5875072f6ac5b571ad2fcb87af9c1f2a",
                "parameters": [
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter conferences starting on or after this date.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter conferences starting on or before this date.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conference statistics retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum",
                "description": "Enter your Sanctum token."
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}